lsp: use diagnostic options properly
Use the diagnostics level to determine whether to store the diagnostics
with the buffer and the diagnostic highlight options to determine
whether to highlight the relevent text.

This will allow users to toggle highlighting to affect whether text is
highlighted and adjust the level to determine whether there are any
diagnostics stored at all.