fmt: adjust imports after formatting on autosave
Order matters when formatting and adjusting imports, because of gopls'
support for gofumpt. Gofumpt formatting will group all imports that look
like a stdlib package (e.g. there's no '.' in the package path)
together. When the local setting is provided, the only way to get the
local imports grouped separately when gofumpt is used to format is to
format first and then organize imports.

Fixes #3151