Support non-gopls LSP servers
This change adds g:go_gopls_binary option that specifies path to the LSP
binary.

Gopls uses custom format for hover response [1], and other LSPs do
not, this change adds bare-bones support for plain text hover responses
(no markdown support).  :GoDocBrowser is driven by fields in structured
responses and is not supported at all, and results in "could not find
path for doc URL" error.

Because non-gopls LSP server support is pretty much experimetal, the
options are not documented, and the option name uses "gopls" (as opposed
to "lsp") as to not create an impression that an arbitrary LSP server
would work out of the box.

Tested:
  Manually in my environment with our internal LSP server and confirmed that
  :GoDoc, :GoInfo, :GoDef work. Admittedly, :GoDoc and :GoInfo are pretty
  much the same, except the latter shows results in the command prompt,
  and the former opens up a doc window.

  `make` reports:
  All (vim-8.0) tests PASSED
  All (vim-8.2) tests PASSED
  All (nvim) tests PASSED

Issues:

  :GoInfo doesn't produce any output when the result has several lines.

Related: https://github.com/fatih/vim-go/issues/3284

[1] https://github.com/golang/tools/blob/bf6c7f26e9f0fea5b256f9ef6968435dddc5be25/internal/lsp/source/hover.go#L562