lint: adjust handling of golangci-lint output
golangci-lint output has changed yet again. This time they seem to be
using relative paths and some kind of weird comment line when the
argument provided to `golangci-lint` is an absolute path. Adjust the
handling of golangci-lint output to not discard messages that start with
a relative path (i.e ./problem.go instead problem.go) and to discard
messages that are the non-relative path (e.g. problem.go), but with the
useless comment.