tests: disable gopls when a test doesn't need it
Disable gopls in many tests that don't need it.

Move the clearing of options to after the the call to go#lsp#Exit in
scripts/runtest to that calling go#lsp#Exit doesn't trigger a new gopls
process to be started in case the test disabled gopls.

Remove unlet g:go_* statements in individual tests, because the runtest
scripts clears them all.