fillstruct: refactor test test that verifies nothing happens
Remove fillstruct test that verifies nothing happens when trying to
calling go#fillstruct#FillStruct with the cursor not on a struct in a
line with more than one struct. This is an edge case in gopls code, and
since nothing happens, the only effect of the test is essentially to
slow down execution of tests.

Replace that test with another that verifies that when there are two
structs on the same line and the cursor is positioned on the first, then
that is the struct that is filled.

Rename an existing test for clarity given that new similarly named test.