syntax: link goPredefinedIdentifiers to Constant
Link goPredefinedIdentifiers to Constant. This is a non-functional
change, because goPredefinedIdentifier was previously linked to
goBoolean, which in turn is linked to Boolean, which is linked to
Constant by default by Vim. This change effectively spells goBoolean
differently. The highlighting for goPredefinedIdentiifers remains the
same as it was and users can still vary the highlighting of goBoolean
and goPredefinedIdentifiers if they wish.