tests: change directories before deleting
Change out of the dynamically created directory before trying to delete
that directory so that the directory change after deletion does not fail
because the current directory at the time of change does not exist.

This situation is also handled by Vim 8.2.0189, but vim-go has to
support 8.1.2269 for now, so adjust tests so that vim-go is no longer
susceptible to the condition that 8.2.0189 fixes.