debug: avoid cleanup errors
Do not attempt to delete a command based on the first line of output
from ":command GoDebug" trying to stop debugging. The error was causing
the function to be aborted, which was in turn causing Neovim to hang
when trying to quit.

Similarly, swallow exceptions when removing commands and mappings.